+++ src/cdn/azext_cdn/aaz/latest/cdn/endpoint/_create.py
@@ -22,9 +22,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2025-09-01-preview",
+        "version": "2026-04-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cdn/profiles/{}/endpoints/{}", "2025-09-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.cdn/profiles/{}/endpoints/{}", "2026-04-01-preview"],
         ]
     }
 
@@ -1353,7 +1353,7 @@ class Create(AAZCommand):
         health_probe_settings.probe_protocol = AAZStrArg(
             options=["probe-protocol"],
             help="Protocol to use for health probe.",
-            enum={"Grpc": "Grpc", "Http": "Http", "Https": "Https", "NotSet": "NotSet"},
+            enum={"Http": "Http", "Https": "Https", "NotSet": "NotSet"},
         )
         health_probe_settings.probe_request_type = AAZStrArg(
             options=["probe-request-type"],
@@ -1692,7 +1692,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2025-09-01-preview",
+                    "api-version", "2026-04-01-preview",
                     required=True,
                 ),
             }