+++ src/horizondb/azext_horizondb/_help.py
@@ -143,3 +143,86 @@ examples:
   - name: Delete a firewall rule.
     text: az horizondb firewall-rule delete --resource-group exampleresourcegroup --cluster-name examplecluster --name allowclientip
 """
+
+
+helps['horizondb private-endpoint-connection'] = """
+type: group
+short-summary: Manage HorizonDB private endpoint connections.
+"""
+
+
+helps['horizondb private-endpoint-connection list'] = """
+type: command
+short-summary: List private endpoint connections for a HorizonDB cluster.
+examples:
+  - name: List private endpoint connections for a HorizonDB cluster.
+    text: az horizondb private-endpoint-connection list --resource-group exampleresourcegroup --cluster-name examplecluster
+"""
+
+
+helps['horizondb private-endpoint-connection show'] = """
+type: command
+short-summary: Show details of a HorizonDB private endpoint connection.
+examples:
+  - name: Show a private endpoint connection by cluster and connection name.
+    text: az horizondb private-endpoint-connection show --resource-group exampleresourcegroup --cluster-name examplecluster --name exampleconnection
+  - name: Show a private endpoint connection by resource ID.
+    text: az horizondb private-endpoint-connection show --id /subscriptions/{subscriptionId}/resourceGroups/{resourceGroup}/providers/Microsoft.HorizonDb/clusters/{clusterName}/privateEndpointConnections/{connectionName}
+"""
+
+
+helps['horizondb private-endpoint-connection approve'] = """
+type: command
+short-summary: Approve a HorizonDB private endpoint connection.
+examples:
+  - name: Approve a private endpoint connection by cluster and connection name.
+    text: az horizondb private-endpoint-connection approve --resource-group exampleresourcegroup --cluster-name examplecluster --name exampleconnection --description "Approved"
+  - name: Approve a private endpoint connection by resource ID.
+    text: az horizondb private-endpoint-connection approve --id /subscriptions/{subscriptionId}/resourceGroups/{resourceGroup}/providers/Microsoft.HorizonDb/clusters/{clusterName}/privateEndpointConnections/{connectionName} --description "Approved"
+"""
+
+
+helps['horizondb private-endpoint-connection reject'] = """
+type: command
+short-summary: Reject a HorizonDB private endpoint connection.
+examples:
+  - name: Reject a private endpoint connection by cluster and connection name.
+    text: az horizondb private-endpoint-connection reject --resource-group exampleresourcegroup --cluster-name examplecluster --name exampleconnection --description "Rejected"
+  - name: Reject a private endpoint connection by resource ID.
+    text: az horizondb private-endpoint-connection reject --id /subscriptions/{subscriptionId}/resourceGroups/{resourceGroup}/providers/Microsoft.HorizonDb/clusters/{clusterName}/privateEndpointConnections/{connectionName} --description "Rejected"
+"""
+
+
+helps['horizondb private-endpoint-connection delete'] = """
+type: command
+short-summary: Delete a HorizonDB private endpoint connection.
+examples:
+  - name: Delete a private endpoint connection by cluster and connection name.
+    text: az horizondb private-endpoint-connection delete --resource-group exampleresourcegroup --cluster-name examplecluster --name exampleconnection
+  - name: Delete a private endpoint connection by resource ID.
+    text: az horizondb private-endpoint-connection delete --id /subscriptions/{subscriptionId}/resourceGroups/{resourceGroup}/providers/Microsoft.HorizonDb/clusters/{clusterName}/privateEndpointConnections/{connectionName}
+"""
+
+
+helps['horizondb private-link-resource'] = """
+type: group
+short-summary: Manage HorizonDB private link resources.
+"""
+
+
+helps['horizondb private-link-resource list'] = """
+type: command
+short-summary: List private link resources for a HorizonDB cluster.
+examples:
+  - name: List private link resources for a HorizonDB cluster.
+    text: az horizondb private-link-resource list --resource-group exampleresourcegroup --cluster-name examplecluster
+"""
+
+
+helps['horizondb private-link-resource show'] = """
+type: command
+short-summary: Show details of a HorizonDB private link resource.
+examples:
+  - name: Show a private link resource by group name.
+    text: az horizondb private-link-resource show --resource-group exampleresourcegroup --cluster-name examplecluster --group-name DefaultPool
+"""